update additional content styling
